lace-platform: Promote linux loader to a real module file
Replace the inline `pub mod linux { pub use super::p::linux::*; }` in
the crate root and the inline mock implementation in `mock/mod.rs`
with proper module files: a top-level `linux.rs` re-exporting the
active platform's `boot_linux` / `BootLinuxError`, and `mock/linux.rs`
holding the mock placeholder. No behavioural change.
