+++ lace-platform/src/lib.rs
@@ -97,6 +97,4 @@ pub mod hwid;
 // Unified filesystem module re-exporting common types and platform-specific functions
 pub mod fs;
 
-pub mod linux {
-    pub use super::p::linux::{BootLinuxError, boot_linux};
-}
+pub mod linux;
