+++ lace-platform/src/mock/mod.rs
@@ -6,6 +6,7 @@
 pub mod console;
 pub mod fs;
 pub mod hwid;
+pub mod linux;
 pub mod mem;
 pub mod tpm2;
 
@@ -16,19 +17,3 @@ use lace_util::Display;
 pub struct Error;
 
 impl std::error::Error for Error {}
-
-pub mod linux {
-    use lace_util::Display;
-
-    #[derive(Debug, Display)]
-    #[display("mock platform boot linux error")]
-    pub struct BootLinuxError;
-
-    pub fn boot_linux(
-        _kernel: &[u8],
-        _initrd: Option<&[u8]>,
-        _cmdline: Option<&str>,
-    ) -> Result<(), BootLinuxError> {
-        todo!()
-    }
-}
